Clean up some `phpcs:ignore` comments and hook parameters
This removes any `1`s passed as the `$accepted_args` parameter from
hooks, since it's the default value.

It also adds verbose `phpcs:ignore` comments throughtout
`council-meetings.php` so we're not ignoring all standards for the
lines in question.

Lastly, this shortens one function name and updates the hook it is
called on.