Generate a post when a Sugar Calendar event is saved
This uses the from the event post to create a post that will redirect
to the event post. Yeah, it's kind of silly, but this feature is being
requested so that events will be included in regular post archive pages
and in blocks that consume feeds of the regular post type.

If a post has already been generated and the event post is updated,
the previously generated post is updated as well.