Use `wp_date()` vs `date()` in `sc_events_*` shortcode queries
The start and end dates attached to events are stored in local time,
which is currently PDT for pcouncil.org. Sugar Calendar originally
used `date()` to get the current day, but this returns the server
time instead, which can be on a different day than local time.

Switching to `wp_date()` builds a date that accounts for the local
timezone and should provide a more relevant query argument.