Update custom sticky input `name` attributes
There is a condition in the `edit_post` function of
`wp-admin/includes/post.php` that checks for `$post_data['sticky']`,
and if it's empty, `unstick_post( $post_ID );` is called.
This is why custom handling to stick posts when `save_post`
runs did not work. :facepalm:

By updating the name attribute, we can just leverage this logic
already in place to handle saving for quick edits and post types
without block editor support.

Testing does indicate that custom handling for bulk editing is
still needed, so at least all that wasn't a waste of time.